requirements.txt
git+ssh://git@github.com/isu-avista/data.git
pyyaml==5.3.1
flask==1.1.2
SQLAlchemy==1.3.20
Flask-SQLAlchemy==2.4.4
Flask-Migrate==2.5.3
python-dotenv==0.15.0